Do not set server uid if effective user id is not super-user
The documentation to the user configuration property
specifies that it shall be ineffective if the effective
user id of xinetd is not super-user. Xinetd should
typically fail to change the user id of a server process
that it spawns if itself does not run as super-user.

Hence, this patch aligns the implementation with the
documented behavior and allows to run xinetd as
unprivileged user.