sio: replace sioexpand with plain realloc
This function is basically realloc with undefined behavior.  The only
time it's called with non-zero (i.e. true) is_static, it's because the
area pointer is NULL.  In that case, malloc is used to allocate memory
before calling memcpy to sync area to new_area -- but area is always
NULL, and old_size is always 0.  This is undefined behavior according
to the C standards.

If we delete the useless memcpy, we're left with a plain malloc, and
that's what realloc does when it's passed a NULL pointer.  Which is
what the other code path in this func does when is_static is 0.  So
delete this entire func and replace callers with realloc directly.

It's simpler, smaller, and has no undefined behavior.